.ipynb_checkpoints/requirements-checkpoint.txt
anyio==4.4.0
python-box==7.2.0
asgi-correlation-id==4.3.2
fastapi==0.112.2
loguru==0.7.2
pydantic==2.8.2
qdrant-client==1.11.1
redis==5.0.8
transformers==4.44.2
uvicorn==0.30.6
vllm==0.5.5
FlagEmbedding==1.2.11
llama-index==0.11.3
llama-index-core==0.11.3
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-vllm==0.2.0
llama-index-postprocessor-flag-embedding-reranker==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.2.0
llama-index-vector-stores-qdrant==0.3.0